# RFC: Optimize for small strings to close gap with FNV