--- a/Cargo.toml
+++ b/Cargo.toml
@@ -15,3 +15,7 @@ rand = ["dep:rand", "std"]
 
 [dependencies]
 rand = { version = "0.8", optional = true }
+
+[dev-dependencies]
+fnv = "1.0"
+rand = "0.8"