--- a/Cargo.toml
+++ b/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rustc-hash"
-version = "1.1.0"
+version = "2.0.0"
 authors = ["The Rust Project Developers"]
 description = "speedy, non-cryptographic hash used in rustc"
 license = "Apache-2.0/MIT"