--- a/src/lib.rs
+++ b/src/lib.rs
@@ -104,29 +104,35 @@ impl FxHasher {
 impl Hasher for FxHasher {
     #[inline]
     fn write(&mut self, mut bytes: &[u8]) {
-        #[cfg(target_pointer_width = "32")]
-        let read_usize = |bytes: &[u8]| u32::from_ne_bytes(bytes[..4].try_into().unwrap());
-        #[cfg(target_pointer_width = "64")]
-        let read_usize = |bytes: &[u8]| u64::from_ne_bytes(bytes[..8].try_into().unwrap());
+        const SIZE: usize = {
+            let size = size_of::<usize>();
+            assert!(1 <= size && size <= 8);
+            size
+        };
 
-        let mut hash = FxHasher { hash: self.hash };
-        assert!(size_of::<usize>() <= 8);
-        while bytes.len() >= size_of::<usize>() {
-            hash.add_to_hash(read_usize(bytes) as usize);
-            bytes = &bytes[size_of::<usize>()..];
+        let mut hash = self.clone();
+
+        while bytes.len() >= SIZE {
+            let (chunk, rest) = bytes.split_at(SIZE);
+            hash.add_to_hash(usize::from_ne_bytes(chunk.try_into().unwrap()));
+            bytes = rest;
         }
-        if (size_of::<usize>() > 4) && (bytes.len() >= 4) {
-            hash.add_to_hash(u32::from_ne_bytes(bytes[..4].try_into().unwrap()) as usize);
-            bytes = &bytes[4..];
+
+        if SIZE > 4 && bytes.len() >= 4 {
+            let (chunk, rest) = bytes.split_at(4);
+            hash.add_to_hash(u32::from_ne_bytes(chunk.try_into().unwrap()) as usize);
+            bytes = rest;
         }
-        if (size_of::<usize>() > 2) && bytes.len() >= 2 {
-            hash.add_to_hash(u16::from_ne_bytes(bytes[..2].try_into().unwrap()) as usize);
-            bytes = &bytes[2..];
+        if SIZE > 2 && bytes.len() >= 2 {
+            let (chunk, rest) = bytes.split_at(2);
+            hash.add_to_hash(u16::from_ne_bytes(chunk.try_into().unwrap()) as usize);
+            bytes = rest;
         }
-        if (size_of::<usize>() > 1) && bytes.len() >= 1 {
+        if SIZE > 1 && bytes.len() >= 1 {
             hash.add_to_hash(bytes[0] as usize);
         }
-        self.hash = hash.hash;
+
+        *self = hash;
     }
 
     #[inline]