@Nilstrieb 2024-01-18
Can you make a PR to https://github.com/rust-lang/rust/ patching rustc-hash to point to your branch? I can then start a rustc benchmark, as that's the benchmark we care about. If it doesn't regress, this seems okay.