@adamreichold 2024-01-19
The rustc benchmarks seem to indicate that the extra rounds for the tail bytes are critical for quality and cannot be merged, i.e. just scheduling changes will not make fxhash competitive with FNV for three-byte strings.